# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Atlas - Project-based technical learning platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /domains/{slug}:
    get:
      operationId: getDomain
      tags: [domains]
      summary: Get domain details with project summary
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Domain detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainDetail"
...
  /modules:
    get:
      operationId: listModules
      tags: [modules]
      summary: List learning modules
      parameters:
        - name: domainSlug
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [python_mastery, sql_mastery, concepts]
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
  /modules/{moduleId}:
    get:
      operationId: getModule
      tags: [modules]
      summary: Get module with lessons
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Module detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleDetail"
  /user/modules/{moduleId}/lessons/{lessonId}/complete:
    post:
      operationId: completeLesson
      tags: [lessons]
      summary: Mark a lesson as complete
      security:
        - clerkAuth: []
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lesson completion recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonCompletion"
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai-tutor]
      summary: Stream AI tutor response (SSE)
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatBody"
      responses:
        "200":
          description: Streaming AI response
          content:
            text/event-stream:
              schema:
                type: string
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ai/chat/history:
    get:
      operationId: getAiChatHistory
      tags: [ai-tutor]
      summary: Get AI chat history for a context
      security:
        - clerkAuth: []
      parameters:
        - name: contextType
          in: query
          required: true
          schema:
            type: string
            enum: [project, module, general]
        - name: contextId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiChatMessage"
...
  /billing/subscription:
    get:
      operationId: getSubscription
      tags: [billing]
      summary: Get user's current subscription
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Current subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /billing/checkout:
    post:
      operationId: createCheckout
      tags: [billing]
      summary: Create a Stripe checkout session
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutBody"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
  /billing/portal:
    post:
      operationId: createBillingPortal
      tags: [billing]
      summary: Create a Stripe billing portal session
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Portal URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalResponse"
  /webhooks/stripe:
    post:
      operationId: stripeWebhook
      tags: [webhooks]
      summary: Stripe webhook handler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
...
  /execute/python:
    post:
      operationId: executePython
      tags: [steps]
      summary: Execute Python code via Piston API
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteCodeBody"
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecuteCodeResult"
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    WebhookResponse:
      type: object
      properties:
        received:
          type: boolean
      required:
        - received
    Domain:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        color:
          type: string
        status:
          type: string
          enum: [active, coming_soon, waitlist]
        projectCount:
          type: integer
        enrolledCount:
          type: integer
      required:
        - id
        - slug
        - name
        - description
        - icon
        - color
        - status
        - projectCount
        - enrolledCount
    DomainDetail:
      allOf:
        - $ref: "#/components/schemas/Domain"
        - type: object
          properties:
            projects:
              type: array
              items:
                $ref: "#/components/schemas/ProjectSummary"
    ProjectSummary:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced, expert]
        tier:
          type: string
          enum: [free, pro]
        xpReward:
          type: integer
        estimatedHours:
          type: number
        stepCount:
          type: integer
        enrolledCount:
          type: integer
        completionRate:
          type: number
        tags:
          type: array
          items:
            type: string
        position:
          type: integer
      required:
        - id
        - slug
        - title
        - description
        - difficulty
        - tier
        - xpReward
        - estimatedHours
        - stepCount
        - enrolledCount
        - completionRate
        - tags
        - position
...
    Module:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [python_mastery, sql_mastery, concepts]
        domainSlug:
          type: string
        lessonCount:
          type: integer
        estimatedHours:
          type: number
        tier:
          type: string
          enum: [free, pro]
        position:
          type: integer
      required:
        - id
        - slug
        - title
        - description
        - type
        - lessonCount
        - estimatedHours
        - tier
        - position
    Lesson:
      type: object
      properties:
        id:
          type: string
        moduleId:
          type: string
        title:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [reading, video, interactive, quiz]
        position:
          type: integer
        estimatedMinutes:
          type: integer
        xpReward:
          type: integer
      required:
        - id
        - moduleId
        - title
        - content
        - type
        - position
        - estimatedMinutes
        - xpReward
    ModuleDetail:
      allOf:
        - $ref: "#/components/schemas/Module"
        - type: object
          properties:
            lessons:
              type: array
              items:
                $ref: "#/components/schemas/Lesson"
            learningObjectives:
              type: array
              items:
                type: string
    LessonCompletion:
      type: object
      properties:
        lessonId:
          type: string
        userId:
          type: string
        completedAt:
          type: string
          format: date-time
        xpEarned:
          type: integer
      required:
        - lessonId
        - userId
        - completedAt
        - xpEarned
    AiChatBody:
      type: object
      properties:
        message:
          type: string
        contextType:
          type: string
          enum: [project, module, general]
        contextId:
          type: string
        conversationId:
          type: string
        currentCode:
          type: string
        stepId:
          type: string
      required:
        - message
        - contextType
    AiChatMessage:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
...
    Subscription:
      type: object
      properties:
        tier:
          type: string
          enum: [free, pro]
        status:
          type: string
          enum: [active, canceled, past_due, trialing, none]
        currentPeriodEnd:
          type: string
          format: date-time
        cancelAtPeriodEnd:
          type: boolean
        stripeCustomerId:
          type: string
      required:
        - tier
        - status
    CreateCheckoutBody:
      type: object
      properties:
        priceId:
          type: string
        billingInterval:
          type: string
          enum: [monthly, annual]
        successUrl:
          type: string
        cancelUrl:
          type: string
      required:
        - priceId
        - billingInterval
    CheckoutResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    PortalResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
...
    ExecuteCodeBody:
      type: object
      properties:
        code:
          type: string
        language:
          type: string
          enum: [python]
        stdin:
          type: string
        timeoutSeconds:
          type: integer
          default: 10
      required:
        - code
        - language
    ExecuteCodeResult:
      type: object
      properties:
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: integer
        executionTimeMs:
          type: integer
        timedOut:
          type: boolean
      required:
        - stdout
        - stderr
        - exitCode
        - timedOut