# lib/api-zod/src/generated/types/domain.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Atlas - Project-based technical learning platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DomainStatus } from "./domainStatus";

export interface Domain {
  id: string;
  slug: string;
  name: string;
  description: string;
  icon: string;
  color: string;
  status: DomainStatus;
  projectCount: number;
  enrolledCount: number;
}